airflow-server/requirements.txt
pandas
evidently
mlflow
torch
torchvision
transformers
prophet
apache-airflow[postgres]
apache-airflow[amazon]
psycopg2-binary
